Allow to build windows binaries (#61)
* Allow to build windows binaries


* Resolve build errores


* Add validation to fail when renewal signal is used on windows, and update README


---------
