+++ .go-version
@@ -1,1 +1,1 @@
-1.19.5
+1.20.1
