+++ helper_windows.conf
@@ -1,0 +1,11 @@
+agentAddress = "spire-agent\\public\\api"
+cmd = ""
+cmdArgs = ""
+certDir = "certs"
+svidFileName = "svid.pem"
+svidKeyFileName = "svid_key.pem"
+svidBundleFileName = "svid_bundle.pem"
+# Add CA with intermediates into Bundle file instead of SVID file,
+# it is the expected behavior in some scenarios like MySQL.
+# Default: false
+# addIntermediatesToBundle = false
